Lots of stuff 3

+++ src/graphics.c
@@ -4,5 +4,5 @@
 #include <string.h>
 
 int middleTextX(int x, char *text) {
-	return (int)round(x / 2.0 - strlen(text) / 2.0);
+	return (int)round(x / 2.0 - (double)strlen(text) / 2.0);
 }
